requirements.txt
annotated-types==0.7.0
anyio==4.7.0
apturl==0.5.2
argcomplete==1.8.1
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.2.2
fasteners==0.14.1
Flask==3.1.0
future==0.18.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==3.2.0
keyring==23.5.0
langchain-core==0.3.28
langsmith==0.2.6
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
livereload==2.6.3
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==3.0.2
mkdocs==1.1.2
monotonic==1.6
more-itertools==8.10.0
msgpack==1.1.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
orjson==3.10.12
packaging==24.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pip-autoremove==0.10.0
pipx==1.0.0
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytesseract==0.3.13
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==1.0.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
systemd-python==234
tenacity==9.0.0
texttable==1.6.4
tornado==6.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
userpath==1.8.0
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
yt-dlp==2025.1.26
zipp==1.0.0